fix reversed ipAddress on little endian architectures
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@12734 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/ip-forward-mib/ipCidrRouteTable/ipCidrRouteTable.h
@@ -96,12 +96,12 @@ extern          "C" {
         /*
          * ipCidrRouteDest(1)/IPADDR/ASN_IPADDRESS/u_long(u_long)//l/A/w/e/r/d/h
          */
-        u_long          ipCidrRouteDest;
+        uint32_t        ipCidrRouteDest;
 
         /*
          * ipCidrRouteMask(2)/IPADDR/ASN_IPADDRESS/u_long(u_long)//l/A/w/e/r/d/h
          */
-        u_long          ipCidrRouteMask;
+        uint32_t        ipCidrRouteMask;
 
         /*
          * ipCidrRouteTos(3)/INTEGER32/ASN_INTEGER/long(long)//l/A/w/e/R/d/h
@@ -111,7 +111,7 @@ extern          "C" {
         /*
          * ipCidrRouteNextHop(4)/IPADDR/ASN_IPADDRESS/u_long(u_long)//l/A/w/e/r/d/h
          */
-        u_long          ipCidrRouteNextHop;
+        uint32_t        ipCidrRouteNextHop;
 
 
     } ipCidrRouteTable_mib_index;
